releaseagent: separate go-images server lifecycle
